Git: move the main git logic into the `Git` commands facade
To separate logic from specific actions implementation. This will help with adding different entry points with similar logic across different action groups.

Also adjust action class names a little bit to better reflect their logic.
