+++ resources/messages/MyBundle.properties
@@ -19,7 +19,7 @@ rebase.successful=Rebase successful
 reset.successful=Reset successful
 updating.project=Updating project
 
-action.GitResetToGreenMasterAction.text=Reset 'master' to Latest Green Commit
-action.GitResetToGreenMasterAction.description=Resets the 'master' branch state to the latest successfully built commit according to TeamCity. Doesn't affect or change the current branch.
-action.GitRebaseOnGreenMasterAction.text=Rebase {0} onto the Latest Green 'master'
-action.GitRebaseOnGreenMasterAction.description=Rebases the current branch on the latest successfully built commit in 'master' branch according to TeamCity.
+action.GitResetMasterToGreenCommitAction.text=Reset 'master' to Latest Green Commit
+action.GitResetMasterToGreenCommitAction.description=Resets the 'master' branch state to the latest successfully built commit according to TeamCity. Doesn't affect or change the current branch.
+action.GitRebaseOnGreenMasterCommitAction.text=Rebase {0} onto the Latest Green 'master'
+action.GitRebaseOnGreenMasterCommitAction.description=Rebases the current branch on the latest successfully built commit in 'master' branch according to TeamCity.
