+++ src/org/jetbrains/kotlin/test/helper/actions/git/KotlinMasterAction.kt
@@ -8,20 +8,9 @@ import git4idea.actions.branch.GitSingleBranchAction
 import git4idea.repo.GitRemote
 import git4idea.repo.GitRepository
 import kotlinx.coroutines.launch
+import org.jetbrains.kotlin.test.helper.git.Git
 
 abstract class KotlinMasterAction : GitSingleBranchAction() {
-    companion object {
-        protected const val MAIN_BRANCH = "master"
-        protected const val DEFAULT_ORIGIN = "origin"
-
-        private val KOTLIN_PUBLIC_MONOREPO_REMOTES = setOf(
-            "git@github.com:JetBrains/kotlin.git",
-            "github.com/JetBrains/kotlin.git",
-            "git@git.jetbrains.team/kt/kotlin.git",
-            "git.jetbrains.team/kt/kotlin.git",
-        )
-    }
-
     protected fun chooseRepository(repositories: List<GitRepository>): GitRepository? = repositories.firstOrNull()
 
     override val disabledForRemote = true
@@ -29,7 +18,7 @@ abstract class KotlinMasterAction : GitSingleBranchAction() {
     override fun isEnabledForRef(ref: GitBranch, repositories: List<GitRepository>) =
         repositories.all { it.isOnBranch }
             && chooseRepository(repositories)?.isPublicKotlinMonorepo() == true
-            && ref.name == MAIN_BRANCH
+            && ref.name == Git.MAIN_BRANCH
             && super.isEnabledForRef(ref, repositories)
 
     override fun actionPerformed(
@@ -50,5 +39,5 @@ abstract class KotlinMasterAction : GitSingleBranchAction() {
         remotes.any { it.isPublicKotlinMonorepo() }
 
     private fun GitRemote.isPublicKotlinMonorepo() =
-        this.urls.any { it.removePrefix("ssh://").removePrefix("https://") in KOTLIN_PUBLIC_MONOREPO_REMOTES }
+        this.urls.any { it.removePrefix("ssh://").removePrefix("https://") in Git.KOTLIN_MONOREPO_REMOTES }
 }
